+++ tests/gbe/test_merge_profile_data_view.py
@@ -10,7 +10,6 @@ from tests.functions.gbe_functions import (
 )
 from gbetext import warn_user_merge_delete
 from gbe.models import Profile
-from gbe.forms import BidBioMergeForm
 
 
 class TestMergeProfileData(TestCase):
@@ -63,17 +62,7 @@ class TestMergeProfileData(TestCase):
         response = self.client.get(self.url)
         self.assertRedirects(response, reverse('home', urlconf="gbe.urls"))
 
-    def test_bad_form_setup(self):
-        msg = "no message"
-        try:
-            form = BidBioMergeForm(initial={"stuff": "random"})
-        except Exception as e:
-            msg = e.args[0]
-        self.assertEqual(msg, "Intial with two profiles are required")
- 
-
     def test_get_form(self):
-        # should exclude selected profile, and user's own profile
         login_as(self.privileged_user, self)
         response = self.client.get(self.url)
         self.assertContains(response, 'Merge Users - Verify Info')
